nothinghide/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nothinghide"
version = "1.0.0"
description = "A security tool to check public exposure of email and password using lawful sources"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "NothingHide Team"}
]
keywords = [
    "security",
    "privacy",
    "breach-check",
    "password-security",
    "hibp",
    "cli",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
]

[project.scripts]
nothinghide = "app.main:app"

[project.urls]
Homepage = "https://github.com/nothinghide/nothinghide"
Documentation = "https://github.com/nothinghide/nothinghide#readme"
Repository = "https://github.com/nothinghide/nothinghide"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]
